requirements.txt
colorama==0.4.6
numpy==1.23.4
opencv_python==4.6.0.66
pyserial==3.5
